+++ srcs/exp/lexer_exp/lexer_tilde_math_exp.c
@@ -6,37 +6,37 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/04 20:46:37 by guvillat          #+#    #+#             */
-/*   Updated: 2020/02/07 05:35:14 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/08 19:07:36 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "sh42.h"
 
-char		*exp_get_tilde(const char *string, int len)
+char	*exp_get_tilde(const char *string, size_t len)
 {
 	(void)string;
 	(void)len;
 	return (ft_strdup("~"));
 }
 
-char		*exp_get_tildep(const char *string, int len)
+char	*exp_get_tildep(const char *string, size_t len)
 {
 	(void)string;
 	(void)len;
 	return (ft_strdup("~+"));
 }
 
-char		*exp_get_tildem(const char *string, int len)
+char	*exp_get_tildem(const char *string, size_t len)
 {
 	(void)string;
 	(void)len;
 	return (ft_strdup("~-"));
 }
 
-char		*exp_get_hook(const char *string, int len)
+char	*exp_get_hook(const char *string, size_t len)
 {
 	char	*str;
-	int		index;
+	size_t	index;
 
 	str = NULL;
 	index = 0;
@@ -55,10 +55,10 @@ char		*exp_get_hook(const char *string, int len)
 	return (NULL);
 }
 
-char		*exp_get_dbparen(const char *string, int len)
+char	*exp_get_dbparen(const char *string, size_t len)
 {
 	char	*str;
-	int		index;
+	size_t	index;
 
 	str = NULL;
 	index = 0;
