+++ srcs/exp/tilde_cmds_math_exp.c
@@ -6,29 +6,29 @@
 /*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/14 16:16:13 by guvillat          #+#    #+#             */
-/*   Updated: 2020/02/07 06:26:54 by arsciand         ###   ########.fr       */
+/*   Updated: 2020/03/08 18:35:20 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "sh42.h"
 
-char	*no_exp(const char *data, t_core *shell)
+char		*no_exp(const char *data, t_core *shell)
 {
 	(void)shell;
 	return (ft_strdup(data));
 }
 
-char	*exp_math(const char *data, t_core *shell)
+char		*exp_math(const char *data, t_core *shell)
 {
 	(void)shell;
 	return (ft_strdup(data));
 }
 
-char	*exp_tilde(const char *data, t_core *shell)
+char		*exp_tilde(const char *data, t_core *shell)
 {
 	t_db	*db_tmp;
 	char	*path[3];
-	int		i;
+	size_t	i;
 
 	i = 0;
 	db_tmp = NULL;
@@ -48,13 +48,13 @@ char	*exp_tilde(const char *data, t_core *shell)
 	return (NULL);
 }
 
-char	*exp_cmd_subs(const char *data, t_core *shell)
+char		*exp_cmd_subs(const char *data, t_core *shell)
 {
 	(void)shell;
 	return (ft_strdup(data));
 }
 
-int		quotes_condition(char c, enum e_estate state)
+u_int8_t	quotes_condition(char c, enum e_estate state)
 {
 	if ((c == '\"' && state == E_QUOTE) || (c == '\'' && state == E_DBQUOTE))
 		return (0);
